refactor - 약속 잡기 네이밍 변경
- make plan -> create plan
